# PERF: ArrowExtensionArray comparison methods